+++ src/core/oah_pair.cc
@@ -100,7 +100,7 @@ void OAHPair::ExpireIfNeeded(uint32_t time_now, uint32_t* set_size, size_t* allo
   }
 }
 
-int64_t OAHPair::ReallocIfNeeded(PageUsage* page_usage, bool* realloced) {
+ssize_t OAHPair::ReallocIfNeeded(PageUsage* page_usage, bool* realloced) {
   *realloced = false;
   if (Empty())
     return 0;
@@ -127,7 +127,7 @@ int64_t OAHPair::ReallocIfNeeded(PageUsage* page_usage, bool* realloced) {
   }
 
   *realloced = true;
-  return static_cast<int64_t>(AllocSize()) - static_cast<int64_t>(old_alloc);
+  return static_cast<ssize_t>(AllocSize()) - static_cast<ssize_t>(old_alloc);
 }
 
 }  // namespace dfly
