+++ src/core/oah_pair.h
@@ -4,6 +4,8 @@
 
 #pragma once
 
+#include <sys/types.h>
+
 #include <cassert>
 #include <cstdint>
 #include <cstring>
@@ -85,7 +87,7 @@ class OAHPair {
     SetTaggedPtr(GetTaggedPtr() | shifted_ext_hash);
   }
 
-  int64_t ReallocIfNeeded(PageUsage* page_usage, bool* realloced);
+  ssize_t ReallocIfNeeded(PageUsage* page_usage, bool* realloced);
 
   char* Raw() const {
     return (char*)(GetTaggedPtr() & ~oah::kTagMask);
